refactor: student 패키지 분리
