+++ src/main/java/in/koreatech/koin/admin/user/service/AdminUserService.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -27,10 +26,6 @@ import in.koreatech.koin.admin.user.dto.AdminOwnersResponse;
 import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
 import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentsResponse;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshRequest;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshResponse;
 import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
@@ -38,12 +33,11 @@ import in.koreatech.koin.admin.user.dto.AdminsCondition;
 import in.koreatech.koin.admin.user.dto.AdminsResponse;
 import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
 import in.koreatech.koin.admin.user.dto.OwnersCondition;
-import in.koreatech.koin.admin.user.dto.StudentsCondition;
 import in.koreatech.koin.admin.user.model.Admin;
 import in.koreatech.koin.admin.user.repository.AdminOwnerRepository;
 import in.koreatech.koin.admin.user.repository.AdminOwnerShopRedisRepository;
 import in.koreatech.koin.admin.user.repository.AdminRepository;
-import in.koreatech.koin.admin.user.repository.AdminStudentRepository;
+import in.koreatech.koin.admin.student.repository.AdminStudentRepository;
 import in.koreatech.koin.admin.user.repository.AdminTokenRepository;
 import in.koreatech.koin.admin.user.repository.AdminUserRepository;
 import in.koreatech.koin.domain.owner.model.Owner;
@@ -51,12 +45,10 @@ import in.koreatech.koin.domain.owner.model.OwnerIncludingShop;
 import in.koreatech.koin.domain.owner.model.OwnerShop;
 import in.koreatech.koin.domain.shop.model.shop.Shop;
 import in.koreatech.koin.domain.student.exception.StudentDepartmentNotValidException;
-import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.domain.student.model.StudentDepartment;
 import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
 import in.koreatech.koin.domain.user.model.User;
-import in.koreatech.koin.domain.user.model.UserGender;
 import in.koreatech.koin.domain.user.model.UserToken;
 import in.koreatech.koin.domain.user.model.UserType;
 import in.koreatech.koin.global.auth.JwtProvider;
@@ -81,16 +73,6 @@ public class AdminUserService {
     private final PasswordEncoder passwordEncoder;
     private final AdminTokenRepository adminTokenRepository;
     private final AdminRepository adminRepository;
-
-    public AdminStudentsResponse getStudents(StudentsCondition studentsCondition) {
-        Integer totalStudents = adminStudentRepository.findAllStudentCount();
-        Criteria criteria = Criteria.of(studentsCondition.page(), studentsCondition.limit(), totalStudents);
-
-        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit());
-        Page<Student> studentsPage = adminStudentRepository.findByConditions(studentsCondition, pageRequest);
-
-        return AdminStudentsResponse.from(studentsPage);
-    }
 
     @Transactional
     public AdminResponse createAdmin(CreateAdminRequest request, Integer adminId) {
@@ -246,26 +228,6 @@ public class AdminUserService {
             }
             adminOwnerShopRedisRepository.deleteById(id);
         }
-    }
-
-    public AdminStudentResponse getStudent(Integer userId) {
-        Student student = adminStudentRepository.getById(userId);
-        return AdminStudentResponse.from(student);
-    }
-
-    @Transactional
-    public AdminStudentUpdateResponse updateStudent(Integer id, AdminStudentUpdateRequest adminRequest) {
-        Student student = adminStudentRepository.getById(id);
-        User user = student.getUser();
-        validateNicknameDuplication(adminRequest.nickname(), id);
-        validateDepartmentValid(adminRequest.major());
-        user.update(adminRequest.nickname(), adminRequest.name(),
-            adminRequest.phoneNumber(), UserGender.from(adminRequest.gender()));
-        user.updateStudentPassword(passwordEncoder, adminRequest.password());
-        student.update(adminRequest.studentNumber(), adminRequest.major());
-        adminStudentRepository.save(student);
-
-        return AdminStudentUpdateResponse.from(student);
     }
 
     public AdminNewOwnersResponse getNewOwners(OwnersCondition ownersCondition) {
